+++ lib/express/index.js
@@ -1,6 +1,8 @@
 const http = require('http');
+const url = require('url');
+
 const parsePaths = require('./parsePaths');
-const handler = require('./handler');
+const route = require('./router');
 const verbs = ['get', 'post'];
 
 Router = {};
@@ -14,7 +16,25 @@ Router.listen = function(port, host, callback) {
 };
 
 // Route requests
-Router.handle = handler;
+Router.handle = function handle(req, res) {
+  // parse url & method, get Data, call route
+  console.log('Routing... ... ...');
+  let verb = req.method.toLowerCase();
+  let path = url.parse(req.url).pathname;
+  if (path.length > 1 && path[path.length - 1] === '/') {
+    path = path.slice(0, path.length - 1);
+  }
+  var p = new Promise((resolve, reject) => {
+    if (verb === 'post') {
+      _getData(req, resolve);
+    } else {
+      resolve();
+    }
+  });
+  p.then(() => {
+    route(req, res, path, verb, Router.routes);
+  });
+};
 
 // Build HTTP verb methods
 verbs.forEach(verb => {
